Put signal-dependent control-flow in compute block.
pointbits contains an instance of signal-dependent control-flow: when
the control flow of the circom program depends on the value of a signal.

Since all signals are undefined at circuit-generation time, this is
has undefined semantics at circuit-generation time. However, during
compute time, it is important.

Thus, it should go in a `compute` block.